Extract base project discovery cache
